+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.2.0.0
 
+* Add `dtbSubTransactions` to `TxBody`
+* Add `subTransactionsTxBodyL` method to `DijkstraEraTxBody` class
 * Add `DijkstraTx` type with `DijkstraTx` and `DijkstraSubTx` constructors
 * Add `DijkstraSubTxBody` constructor to `DijkstraTxBodyRaw`
 * Add `TxLevel` argument to `Tx` and `TxBody`